Refactor USE_PROXY
